requirements.txt
amqp==5.1.1
asgiref==3.5.1
async-timeout==4.0.2
autopep8==1.6.0
billiard==3.6.4.0
celery==5.2.7
cffi==1.15.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==37.0.2
Deprecated==1.2.13
Django==4.0.4
django-cors-headers==3.12.0
django-environ==0.9.0
django-filter==22.1
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
drf-jwt==1.19.2
kombu==5.2.4
mysqlclient==2.1.0
packaging==21.3
phonenumbers==8.12.48
Pillow==9.1.1
prompt-toolkit==3.0.29
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
pytz==2022.1
redis==4.3.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
